[ios] Introduce MGLAnnotationView
Add an UIView subclass that can be used as the base class for all
client provided UIViews for annotations. This is being done with
the expectation that it'll be useful in the future to hold additional
information about the view and will help when view reuse is eventually
added for annotations.

This commit also bootstraps the MGLMapView to be able to display
UIViews over the map if provided by the client delegate. For now,
if the delegate provides a UIView then it will be used. If not, the
map view will fall back to the old strategy of using GL annotations
with an image provided by the delegate or a default image if not.

Finally, iosapp now implements the new delegate method to supply
a native view so all annotations look like the simple, blue, semi-trans
parent circle that it currently provides in all cases.